requirements.txt
boto3==1.9.238
botocore==1.12.238
certifi==2019.9.11
chardet==3.0.4
docutils==0.15.2
idna==2.8
jmespath==0.9.4
python-dateutil==2.8.0
PyYAML==5.1.2
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
urllib3==1.25.6
